+++ packages/shopify-app-remix/src/server/boundary/__tests__/error.test.tsx
@@ -13,6 +13,16 @@ describe('Error boundary', () => {
     );
   });
 
+  it('returns a string when handling an ErrorResponseImpl', () => {
+    // WHEN
+    const result = boundary.error(new ErrorResponseImpl());
+
+    // THEN
+    expect(result).toEqual(
+      <div dangerouslySetInnerHTML={{__html: 'Handling response'}} />,
+    );
+  });
+
   it('throws an error when handling an unknown error', () => {
     // WHEN
     const result = () => boundary.error(new Error());
@@ -23,3 +33,4 @@ describe('Error boundary', () => {
 });
 
 class ErrorResponse extends Error {}
+class ErrorResponseImpl extends Error {}