+++ packages/shopify-app-remix/src/server/boundary/error.tsx
@@ -1,5 +1,8 @@
 export function errorBoundary(error: any) {
-  if (error.constructor.name === 'ErrorResponse') {
+  if (
+    error.constructor.name === 'ErrorResponse' ||
+    error.constructor.name === 'ErrorResponseImpl'
+  ) {
     return (
       <div
         dangerouslySetInnerHTML={{__html: error.data || 'Handling response'}}