+++ packages/shopify-app-remix/CHANGELOG.md
@@ -1,5 +1,49 @@
 # @shopify/shopify-app-remix
 
+## 2.1.0
+
+### Minor Changes
+
+- f34eefd: Added v3_authenticatePublic feature flag to remove `authenticate.public(request)`.
+
+  Apps can opt in to the new future at any time, so this is not a breaking change until version 3.
+
+    <details>
+      <summary>See an example</summary>
+
+  Without the `v3_authenticatePublic` future flag the deprecated `authenticate.public(request)` is supported:
+
+  ```ts
+  await authenticate.public.checkout(request);
+  await authenticate.public.appProxy(request);
+
+  // Deprecated.  Use authenticate.public.checkout(request) instead
+  await authenticate.public(request);
+  ```
+
+  With the `v3_authenticatePublic` future flag enabled the deprecated `authenticate.public(request)` is not supported:
+
+  ```ts
+  await authenticate.public.checkout(request);
+  await authenticate.public.appProxy(request);
+  ```
+
+    </details>
+
+## 2.0.2
+
+### Patch Changes
+
+- ee7114a: Fixed the errorBoundary to work with new cases in Remix v2. Thank you @btomaj!
+
+## 2.0.1
+
+### Patch Changes
+
+- 6d12840: Updating dependencies on @shopify/shopify-api
+- Updated dependencies [6d12840]
+  - @shopify/shopify-app-session-storage@2.0.1
+
 ## 2.0.0
 
 ### Major Changes